+++ djangoProject/urls.py
@@ -16,15 +16,16 @@ urlpatterns = [
     path('team/<int:id>', views.team, name='team'),
     path('grappelli/', include('grappelli.urls')),
     path('logout/', views.my_logout, name='logout'),
-    path('password-reset/', auth_views.PasswordResetView.as_view(template_name='password_reset.html'),
+    path('password-reset/', auth_views.PasswordResetView.as_view(template_name='authorization/password_reset.html'),
          name='password_reset'),
-    path('password-reset/done/', auth_views.PasswordResetDoneView.as_view(template_name='password_reset_done.html'),
+    path('password-reset/done/', auth_views.PasswordResetDoneView.as_view(template_name='authorization'
+                                                                                        '/password_reset_done.html'),
          name='password_reset_done'),
     path('password-reset/confirm/<uidb64>/<token>',
-         auth_views.PasswordResetConfirmView.as_view(template_name='password_reset_confirm.html'),
+         auth_views.PasswordResetConfirmView.as_view(template_name='authorization/password_reset_confirm.html'),
          name='password_reset_confirm'),
     path('password-reset-complete/',
-         auth_views.PasswordResetCompleteView.as_view(template_name='password_reset_complete.html'),
+         auth_views.PasswordResetCompleteView.as_view(template_name='authorization/password_reset_complete.html'),
          name='password_reset_complete'),
     path('', views.index, name='home'),
     path('user/', include('users.urls')),
