+++ djangoProject/views.py
@@ -4,7 +4,6 @@ from users.models import House, Teams
 from blog.models import Post
 from django.shortcuts import get_object_or_404
 from django.contrib.auth import logout
-from djangoProject.forms import UrlRequestForm
 
 
 def index(request):
@@ -36,7 +35,7 @@ def house(request, id):
     context = {
         'query_set': house_set
     }
-    return render(request, 'house.html', context=context)
+    return render(request, 'member-list/house.html', context=context)
 
 
 def team(request, id):
@@ -44,7 +43,7 @@ def team(request, id):
     context = {
         'query_set': team_set
     }
-    return render(request, 'team.html', context=context)
+    return render(request, 'member-list/team.html', context=context)
 
 
 def handler404(request, exception):
@@ -82,15 +81,3 @@ def handler403(request, exception):
 def my_logout(request):
     logout(request)
     return redirect('home')
-
-
-def timer(request):
-    if request.POST:
-        form = UrlRequestForm(request.POST)
-        if form.is_valid():
-            url = form.cleaned_data['URL']
-            return render(request, 'timer/timer.html', context={'url': url,'form':form})
-        return render(request, 'timer/timer.html', context={'url': 'https://cuckoo.team/'})
-    else:
-        form = UrlRequestForm()
-        return render(request, 'timer/timer.html', context={'url': 'https://cuckoo.team/','form':form})
