+++ templates/badges/rewards.html
@@ -64,7 +64,7 @@
                                         {% for reward in query %}
                                             <tr>
                                                 <td class="font-weight-bold">
-                                                    {{ reward.timestamp | date:"d-M f a"}}
+                                                    {{ reward.timestamp | date : "d-M f a" }}
                                                 </td>
                                                 <td>
                                                     <div class="media align-items-center">
