--- templates/logs-users.html
@@ -1,57 +1,0 @@
-{% extends 'base.html' %}
-{% load static %}
-{% load crispy_forms_filters %}
-{% load crispy_forms_tags %}
-{% block style %}
-    <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.11.4/themes/smoothness/jquery-ui.min.css"/>
-<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/jquery-ui-timepicker-addon/1.6.3/jquery-ui-timepicker-addon.min.css"/>
-{% endblock style %}
-
-
-{% block content %}
-
-    <main id="main">
-        <div class="breadcrumbs" data-aos="fade-in">
-            <div class="container">
-                <h2>{{ heading }}</h2>
-            </div>
-        </div>
-
-
-    </main>
-
-
-        <section id="about" class="container">
-            <div class="container" data-aos="fade-down">
-                <div class="col-lg-8 offset-lg-2 col-12">
-                    <form method="POST" action="">
-                        {% csrf_token %}
-                        {{ form | crispy }}
-                        <button type="submit" class="btn btn-block get-started-btn-special">Export Data <span class="icofont-arrow-down"></span></button>
-                    </form>
-                </div>
-            </div>
-        </section>
-{% endblock content %}
-
-{% block script %}
-    <script>
-        $(function () {
-            $("#id_beginning").datetimepicker({
-                dateFormat: 'yy-mm-dd',
-                changeMonth: true,
-                changeYear: true,
-                yearRange: "-100:+0",
-            }).attr("autocomplete", "off");
-            $("#id_end").datetimepicker({
-                dateFormat: 'yy-mm-dd',
-                changeMonth: true,
-                changeYear: true,
-                yearRange: "-100:+0",
-            }).attr("autocomplete", "off");
-
-        });
-    </script>
-    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.js" integrity="sha512-uto9mlQzrs59VwILcLiRYeLKPPbS/bT71da/OEBYEwcdNUk8jYIy+D176RYoop1Da+f9mvkYrmj5MCLZWEtQuA==" crossorigin="anonymous"></script>
-    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery-ui-timepicker-addon/1.6.3/jquery-ui-timepicker-addon.min.js" integrity="sha512-s5u/JBtkPg+Ff2WEr49/cJsod95UgLHbC00N/GglqdQuLnYhALncz8ZHiW/LxDRGduijLKzeYb7Aal9h3codZA==" crossorigin="anonymous"></script>
-{% endblock script %}
