--- templates/reward.html
@@ -1,20 +1,0 @@
-{% extends 'base.html' %}
-{% load static %}
-{% load render_table from django_tables2 %}
-{% block content %}
-
-    <main id="main">
-        <div class="breadcrumbs" data-aos="fade-in">
-            <div class="container">
-                <h2>Badges Awarded</h2>
-            </div>
-        </div>
-
-        <section id="schedule" class="section-with-bg text-center">
-            {% render_table object_list %}
-        </section>
-    </main>
-
-
-
-{% endblock content %}
